fix : shift main activity to kotlin with mvp6
