+++ app/src/main/java/org/fossasia/susi/ai/data/contract/IUtilModel.kt
@@ -27,7 +27,7 @@ interface IUtilModel {
     fun putBooleanPref(prefName: String, value: Boolean)
     fun checkMicInput(): Boolean
     fun copyAssetstoSD()
-    fun createSpeechRecognizer(): SpeechRecognizer
     fun decodeImage(previouslyChatImage: String): Drawable
     fun permissionsToGet() : Array<String>
+    fun isArmDevice(): Boolean
 }
